requirements.txt
pygame-ce==2.2.0.dev2
PyTMX==3.31
moderngl==5.7.4
PygameShader==1.0.8